requirements.txt
aiogram==2.15
aiohttp==3.8.0
aiosignal==1.2.0
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.2.0
Babel==2.9.1
billiard==3.6.4.0
bleach==5.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
Django==4.1.2
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.5
django-mptt==0.14.0
django-parler==2.3
django-parler-rest==2.2
django-rest-framework==0.1.0
django-rosetta==0.9.8
django-summernote==0.8.20.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
environs==8.0.0
facebook-sdk==3.1.0
frozenlist==1.2.0
future==0.18.2
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.14.0
multidict==5.2.0
oscrypto==1.3.0
packaging==21.3
Pillow==9.3.0
polib==1.1.1
prompt-toolkit==3.0.33
protobuf==4.21.9
psycopg==3.1.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dotenv==0.19.1
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.3.5
reportlab==3.6.12
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.5.8
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ua-parser==0.16.1
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.12
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.8
yarl==1.7.2
